move traits to tree.rs
* also revamp docs
* remove re-exports that are not needed
